user is now able to use show info and exception is thrown if invalid

+++ src/main/java/nz/ac/auckland/se281/Country.java
@@ -19,7 +19,7 @@ public class Country {
     return continent;
   }
 
-  public double getTaxFees() {
+  public int getTaxFees() {
     return taxFees;
   }
